Prepare v2.3.0

+++ sample/sample_benchmark_coroutine.cpp
@@ -71,14 +71,14 @@ int main(int argc, char *argv[]) {
 
   size_t stack_size = 16 * 1024;
   if (argc > 3) {
-    stack_size = atoi(argv[3]) * 1024;
+    stack_size = static_cast<size_t>(atoi(argv[3]) * 1024);
   }
 
   time_t begin_time = time(nullptr);
   CALC_CLOCK_T begin_clock = CALC_CLOCK_NOW();
 
   // create coroutines
-  co_arr = new copp::coroutine_context_default::ptr_t[max_coroutine_number];
+  co_arr = new copp::coroutine_context_default::ptr_t[static_cast<size_t>(max_coroutine_number)];
 
   time_t end_time = time(nullptr);
   CALC_CLOCK_T end_clock = CALC_CLOCK_NOW();
